Move exec functions to runner
